+++ packages/mobx-state-tree/src/index.ts
@@ -130,10 +130,7 @@ export {
     Instance,
     SnapshotIn,
     SnapshotOut,
-    SnapshotOrInstance,
-    modelAction,
-    modelView,
-    modelState
+    SnapshotOrInstance
 } from "./internal"
 
 export * from "./core/mst-operations"